+++ lib/webview2gtk/webview.vala
@@ -250,6 +250,7 @@ public class WebView : Gtk.Box {
 	}
 
 	~WebView() {
+		Win32Atspi.Bridge.unregister(this);
 		if (attached) {
 			wv2_host_set_permission_handler(host_handle, null, null);
 			wv2_host_set_script_message_handler(host_handle, null, null);